examples/ainews/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.8
aiosignal==1.3.1
airtrain-py==0.1.2
annotated-types==0.7.0
anyio==4.6.0
attrs==24.2.0
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.3.2
chromedriver-autoinstaller==0.6.4
click==8.1.7
cssselect==1.2.0
dataclasses-json==0.6.7
deprecated==1.2.14
dirtyjson==1.0.8
distro==1.9.0
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.16.1
frozenlist==1.4.1
fsspec==2024.9.0
greenlet==3.0.3
h11==0.14.0
html2text==2024.2.26
httpcore==1.0.6
httpx==0.27.2
idna==3.10
jieba3k==0.35.1
jiter==0.5.0
joblib==1.4.2
llama-index-core==0.11.15
llama-index-embeddings-openai==0.2.5
llama-index-readers-web==0.2.3
llamaindex-py-client==0.1.19
lxml==5.3.0
marshmallow==3.22.0
multidict==6.1.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.3
newspaper3k==0.2.8
nltk==3.9.1
numpy==1.26.4
openai==1.51.0
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.3
pillow==10.4.0
playwright==1.47.0
pyarrow==17.0.0
pydantic==2.9.2
pydantic-core==2.23.4
pyee==12.0.0
pysocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.2
pyyaml==6.0.2
regex==2024.9.11
requests==2.32.3
requests-file==2.1.0
selenium==4.25.0
sgmllib3k==1.0.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
spider-client==0.0.27
sqlalchemy==2.0.35
tenacity==8.5.0
tiktoken==0.7.0
tinysegmenter==0.3
tldextract==5.1.2
tqdm==4.66.5
trio==0.26.2
trio-websocket==0.11.1
typing-extensions==4.12.2
typing-inspect==0.9.0
tzdata==2024.2
urllib3==2.2.3
websocket-client==1.8.0
wrapt==1.16.0
wsproto==1.2.0
yarl==1.13.1
